Good news:  All the new entity collision check functions seem to work for 2 entity collisions at least :)

+++ PerspectiveGame/tile.h
@@ -56,7 +56,7 @@ public:
 	glm::fvec3 color;
 	
 	int entityIndices[9];
-	uint_fast8_t entityInfoIndices[9];
+	int entityInfoIndices[9];
 
 	glm::vec2 texCoords[4];
 	bool cornerSafety[4];
